refactor(player): rename classes and clean up
