--- external_projects/bccm_player/lib/src/widgets/bccm_player_controls.dart
@@ -1,11 +1,0 @@
-import 'package:flutter/foundation.dart';
-import 'package:flutter/material.dart';
-
-class BccmPlayerControls extends StatelessWidget {
-  const BccmPlayerControls({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const Placeholder();
-  }
-}
