+++ external_projects/bccm_player/lib/src/widgets/video/video_platform_view.dart
@@ -1,0 +1,124 @@
+import 'package:universal_io/io.dart';
+
+import 'package:flutter/foundation.dart';
+import 'package:flutter/gestures.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart';
+import 'package:flutter/services.dart';
+
+class VideoPlatformView extends StatelessWidget {
+  final String id;
+  final bool showControls;
+
+  const VideoPlatformView({
+    super.key,
+    required this.id,
+    this.showControls = true,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    Widget getPlatformSpecificPlayer() {
+      if (kIsWeb) {
+        return _WebPlayer(parent: this);
+      } else if (Platform.isAndroid) {
+        return _AndroidPlayer(parent: this);
+      } else if (Platform.isIOS) {
+        return _IOSPlayer(parent: this);
+      }
+      return const SizedBox.shrink();
+    }
+
+    return AspectRatio(
+      aspectRatio: 16 / 9,
+      child: getPlatformSpecificPlayer(),
+    );
+  }
+}
+
+class _WebPlayer extends StatelessWidget {
+  const _WebPlayer({
+    required this.parent,
+  });
+
+  final VideoPlatformView parent;
+
+  @override
+  Widget build(BuildContext context) {
+    return HtmlElementView(viewType: 'bccm-player-${parent.id}');
+  }
+}
+
+class _IOSPlayer extends StatelessWidget {
+  const _IOSPlayer({
+    required this.parent,
+  });
+
+  final VideoPlatformView parent;
+
+  @override
+  Widget build(BuildContext context) {
+    return UiKitView(
+      viewType: 'bccm-player',
+      hitTestBehavior: PlatformViewHitTestBehavior.translucent,
+      gestureRecognizers: <Factory<OneSequenceGestureRecognizer>>{
+        Factory<OneSequenceGestureRecognizer>(
+          () => EagerGestureRecognizer(),
+        ),
+      },
+      creationParams: <String, dynamic>{
+        'player_id': parent.id,
+        'show_controls': parent.showControls,
+      },
+      creationParamsCodec: const StandardMessageCodec(),
+    );
+  }
+}
+
+class _AndroidPlayer extends StatelessWidget {
+  const _AndroidPlayer({
+    Key? key,
+    required this.parent,
+  }) : super(key: key);
+
+  final VideoPlatformView parent;
+
+  @override
+  Widget build(BuildContext context) {
+    return PlatformViewLink(
+      viewType: 'bccm-player',
+      surfaceFactory: (context, controller) {
+        debugPrint("viewId ${controller.viewId}");
+        return AndroidViewSurface(
+          controller: controller as AndroidViewController,
+          hitTestBehavior: PlatformViewHitTestBehavior.opaque,
+          gestureRecognizers: {
+            Factory<TapGestureRecognizer>(() => TapGestureRecognizer()),
+            Factory<HorizontalDragGestureRecognizer>(() => HorizontalDragGestureRecognizer()),
+          },
+        );
+      },
+      onCreatePlatformView: (params) {
+        var controller = PlatformViewsService.initExpensiveAndroidView(
+          id: params.id,
+          viewType: 'bccm-player',
+          layoutDirection: TextDirection.ltr,
+          creationParams: <String, dynamic>{
+            'player_id': parent.id,
+            'show_controls': parent.showControls,
+          },
+          creationParamsCodec: const StandardMessageCodec(),
+          onFocus: () {
+            params.onFocusChanged(true);
+          },
+        );
+        controller
+          ..addOnPlatformViewCreatedListener((val) {
+            params.onPlatformViewCreated(val);
+          })
+          ..create();
+        return controller;
+      },
+    );
+  }
+}
